Change size_t in FileBuffer to int

+++ common/src/filesystem/FileBuffer.hpp
@@ -12,7 +12,7 @@ namespace avansync
 
     virtual char* data() = 0;
     virtual size_t size() const = 0;
-    virtual void reserve(size_t capacity) = 0;
+    virtual void reserve(int capacity) = 0;
   };
 
 } // namespace avansync
